backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dateutil>=2.9.0
vaderSentiment>=3.3.2